func/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-identity
openai
azure-monitor-opentelemetry
azure-monitor-events-extension
tiktoken
urllib3
starlette